pyproject.toml
[tool.poetry]
name = "moonbot"
version = "0.1.0"
description = ""
authors = ["Dmitrii Belov <dmitrii.belov99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
